+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 
 ## Since v1.9
-Fundamental change to Pieces (now using a list of vectors for drawn edges instead of a single vector for join edge).
+
+- Possibly breaking change:
+Redefined Pieces to use a list of vectors for the drawn edges instead of a single vector for the join edge.
 VPatch and Patch now transform correctly before and after drawing (linear transformations).
-Previously only rotations, translations, and uniform scaling worked before drawing.
+(Previously only rotations, translations, and uniform scaling worked before drawing.)
+The same changes have been made to P3_Piece in TileLibP3 so P3_Patch will transform correctly.
+These no longer use angles in their definition: drawnEdges, drawnEdgesP3, decompPieceP3toP2, decompPieceP2toP3
+
+- The following do rely on angles so will not work as expected after a transformation that does not preserve angles: wholeTileEdges, leftFillPieceDK.
 
 ## version 1.9